src/main/java/pablosja/lindisfarne/dtos/UserRegistrationDto.java
package pablosja.lindisfarne.dtos;

public class UserRegistrationDto {
    private String username;
    private String encryptedPassword;

    public UserRegistrationDto() {}

    public UserRegistrationDto(String username, String encryptedPassword) {
        this.username = username;
        this.encryptedPassword = encryptedPassword;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getEncryptedPassword() {
        return encryptedPassword;
    }

    public void setEncryptedPassword(String encryptedPassword) {
        this.encryptedPassword = encryptedPassword;
    }
}
